pyproject.toml
[project]
name = "ai-jobs-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "openai>=2.29.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]